requirements.txt
python-dotenv
transformers
torch
bitsandbytes
accelerate
requests
psycopg[binary,pool]>=3.1.18
pgvector>=0.2.5
beautifulsoup4
lxml
openai
playwright
numpy
tiktoken
feedparser 
pdfminer.six
langchain 
langgraph 
langchain-community
